+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/style/standard/StandardStyleConfigurationState.kt
@@ -35,6 +35,7 @@ public class StandardStyleConfigurationState private constructor(
   initialColorTrunks: ColorValue,
   initialColorWater: ColorValue,
   initialDensityPointOfInterestLabels: DoubleValue,
+  initialFuelingStationModePointOfInterestLabels: StringValue,
   initialRoadsBrightness: DoubleValue,
   initialShowAdminBoundaries: BooleanValue,
   initialShowLandmarkIconLabels: BooleanValue,
@@ -65,6 +66,7 @@ public class StandardStyleConfigurationState private constructor(
     initialColorTrunks = ColorValue.INITIAL,
     initialColorWater = ColorValue.INITIAL,
     initialDensityPointOfInterestLabels = DoubleValue.INITIAL,
+    initialFuelingStationModePointOfInterestLabels = StringValue.INITIAL,
     initialRoadsBrightness = DoubleValue.INITIAL,
     initialShowAdminBoundaries = BooleanValue.INITIAL,
     initialShowLandmarkIconLabels = BooleanValue.INITIAL,
@@ -168,6 +170,11 @@ public class StandardStyleConfigurationState private constructor(
   public var densityPointOfInterestLabels: DoubleValue by mutableStateOf(initialDensityPointOfInterestLabels)
 
   /**
+   * Control the visibility of fuel and electric charging station POI labels. Default displays both types.
+   */
+  public var fuelingStationModePointOfInterestLabels: StringValue by mutableStateOf(initialFuelingStationModePointOfInterestLabels)
+
+  /**
    * Control how bright roads appear in dark styles.
    */
   public var roadsBrightness: DoubleValue by mutableStateOf(initialRoadsBrightness)
@@ -218,6 +225,7 @@ public class StandardStyleConfigurationState private constructor(
     internal const val CONFIG_COLOR_TRUNKS = "colorTrunks"
     internal const val CONFIG_COLOR_WATER = "colorWater"
     internal const val CONFIG_DENSITY_POINT_OF_INTEREST_LABELS = "densityPointOfInterestLabels"
+    internal const val CONFIG_FUELING_STATION_MODE_POINT_OF_INTEREST_LABELS = "fuelingStationModePointOfInterestLabels"
     internal const val CONFIG_ROADS_BRIGHTNESS = "roadsBrightness"
     internal const val CONFIG_SHOW_ADMIN_BOUNDARIES = "showAdminBoundaries"
     internal const val CONFIG_SHOW_LANDMARK_ICON_LABELS = "showLandmarkIconLabels"